model: give each parametric filter its own module
filters.py held all three filter heads plus the pieces they share, which made
the file a reader opens to understand one filter 711 lines long.

Each filter is now in the module named after it - cubic.py, graduated.py,
elliptical.py - with the coordinate-grid cache, the straight-through
binarisation and the gate helper in filtercommon.py. filters.py stays as the
import surface, so filters.CubicFilter and the rest keep resolving.

Class bodies are moved unchanged; no module hierarchy changes, so state dicts
and existing checkpoints are unaffected. 47 tests pass.
